feat: Enhance media playback and UI components
- Implemented pending play functionality in MediaPlayerService to handle playback requests before media is ready.
- Updated PlayerViewModel to remove unnecessary delay before playing media.
- Refactored SoundWaveIndicator to use asynchronous scaling methods for smoother animations.
- Redesigned EqualizerPage to utilize a custom VerticalSlider control for better user interaction.
- Improved PlayerPage layout for fullscreen and normal modes, ensuring proper media element positioning.
- Added new SVG icons for tabs and settings to enhance UI consistency.
- Created AndroidLibraryScanner service for scanning audio and video files from device storage.
- Removed deprecated strings.xml and added necessary resources for Android.

+++ Tips Player/Platforms/Android/MediaPlaybackService.cs
@@ -80,8 +80,8 @@ public class MediaPlaybackService : Service
         _mediaSession = new MediaSession(this, "TipsPlayerSession");
         _mediaSession.SetCallback(new MediaSessionCallback());
         _mediaSession.SetFlags(
-            (int)(MediaSessionFlags.HandlesMediaButtons |
-                  MediaSessionFlags.HandlesTransportControls));
+            MediaSessionFlags.HandlesMediaButtons |
+            MediaSessionFlags.HandlesTransportControls);
         _mediaSession.Active = true;
     }
